# swagger/swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:9090
...
paths:
...
  /api/v1/education-institution:
    post:
      operationId: createInstitution
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInstitutionCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EducationInstitutionDto"
          description: Institution created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new education institution
      tags:
      - Education Institution
    put:
      operationId: updateInstitution
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInstitutionUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EducationInstitutionDto"
          description: Institution updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing education institution
      tags:
      - Education Institution
  /api/v1/education-institution/{id}:
    delete:
      operationId: deleteInstitution
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Institution deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete an education institution by ID
      tags:
      - Education Institution
    get:
      operationId: getInstitutionById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EducationInstitutionDto"
          description: Institution found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get education institution by ID
      tags:
      - Education Institution
  /api/v1/education/{id}:
    delete:
      operationId: deleteEducation
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Education record deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete an education record by ID
      tags:
      - Education
    get:
      operationId: getEducationById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EducationDto"
          description: Education record found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get education record by ID
      tags:
      - Education
  /api/v1/job:
    post:
      operationId: createJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JobDto"
          description: Job record created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new job record
      tags:
      - Job
    put:
      operationId: updateJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JobDto"
          description: Job record updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing job record
      tags:
      - Job
  /api/v1/job/{id}:
    delete:
      operationId: deleteJob
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Job record deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete a job record by ID
      tags:
      - Job
    get:
      operationId: getJobById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JobDto"
          description: Job record found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get a job record by ID
      tags:
      - Job
...
components:
  schemas:
    BadRequestErrorResponse:
      type: object
      description: Bad Request Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "400"
        message:
          type: string
          description: Description of the error
          example: Invalid input provided
...
    CompanyDto:
      type: object
      description: Company Model
      properties:
        id:
          type: integer
          format: int64
          description: ID of the company
        name:
          type: string
          description: Name of the company
        url:
          type: string
          description: Company URL
...
    ConflictErrorResponse:
      type: object
      description: Conflict Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "409"
        message:
          type: string
          description: Description of the error
          example: Data Conflict
...
    EducationDto:
      type: object
      description: Education Record Model
      properties:
        attendanceFormat:
          type: string
          description: Education attendance format
          enum:
          - ON_SITE
          - HYBRID
          - REMOTE
        degreeName:
          type: string
          description: Degree name
        degreeType:
          type: string
          description: Degree type
          enum:
          - UNIVERSITY
          - COLLEGE
          - HIGH_SCHOOL
          - ELEMENTARY_SCHOOL
          - OTHER
        endDate:
          type: string
          format: date-time
          description: End date of the education
        fieldOfStudy:
          type: string
          description: Field of study
        format:
          type: string
          description: "Education format ([rus] ОЧНОЕ, ОЧНО-ЗАОЧНОЕ, ЗАОЧНОЕ)"
          enum:
          - FULL_TIME
          - PART_TIME
          - CORRESPONDENCE
        grade:
          type: string
          description: Grade obtained
        id:
          type: integer
          format: int64
          description: Education ID
        institution:
          $ref: "#/components/schemas/EducationInstitutionDto"
        institutionSubdivision:
          type: string
          description: "Education institution subdivision (Faculty, Institute, etc.)"
        specialization:
          type: string
          description: Specialization
        startDate:
          type: string
          format: date-time
          description: Start date of the education
    EducationInstitutionCreateDto:
      type: object
      description: Education Institution Create Model
      properties:
        name:
          type: string
          description: Name of the education institution
          maxLength: 255
          minLength: 0
      required:
      - name
    EducationInstitutionDto:
      type: object
      description: Education Institution Model
      properties:
        id:
          type: integer
          format: int64
          description: ID of the education institution
        name:
          type: string
          description: Name of the education institution
    EducationInstitutionUpdateDto:
      type: object
      description: Education Institution Update Model
      properties:
        id:
          type: integer
          format: int64
          description: ID of the education institution
        name:
          type: string
          description: Name of the education institution
          maxLength: 255
          minLength: 0
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "500"
        message:
          type: string
          description: Description of the error
          example: Internal Server Error
    JobCreateDto:
      type: object
      description: Job Create Model
      properties:
        attendanceFormat:
          type: string
          description: Job attendance format
          enum:
          - ON_SITE
          - HYBRID
          - REMOTE
        companyId:
          type: integer
          format: int64
          description: ID of the company
        description:
          type: string
          description: Job description
          maxLength: 2000
          minLength: 0
        endDate:
          type: string
          format: date-time
          description: Job end date
        location:
          type: string
          description: Job location
        resumeId:
          type: integer
          format: int64
          description: ID of the resume
        startDate:
          type: string
          format: date-time
          description: Job start date
        title:
          type: string
          description: Job title
          maxLength: 255
          minLength: 0
      required:
      - resumeId
      - title
    JobDto:
      type: object
      description: Job Model
      properties:
        attendanceFormat:
          type: string
          description: Job attendance format
          enum:
          - ON_SITE
          - HYBRID
          - REMOTE
        company:
          $ref: "#/components/schemas/CompanyDto"
        description:
          type: string
          description: Job description
        endDate:
          type: string
          format: date-time
          description: Job end date
        id:
          type: integer
          format: int64
          description: Job ID
        location:
          type: string
          description: Job location
        startDate:
          type: string
          format: date-time
          description: Job start date
        title:
          type: string
          description: Job title
    JobUpdateDto:
      type: object
      description: Job Update Model
      properties:
        attendanceFormat:
          type: string
          description: Job attendance format
          enum:
          - ON_SITE
          - HYBRID
          - REMOTE
        companyId:
          type: integer
          format: int64
          description: Company ID
        description:
          type: string
          description: Job description
          maxLength: 2000
          minLength: 0
        endDate:
          type: string
          format: date-time
          description: Job end date
        id:
          type: integer
          format: int64
          description: Job ID
        location:
          type: string
          description: Job location
        resumeId:
          type: integer
          format: int64
          description: Resume ID
        startDate:
          type: string
          format: date-time
          description: Job start date
        title:
          type: string
          description: Job title
          maxLength: 255
          minLength: 0
...
    NotFoundErrorResponse:
      type: object
      description: Not Found Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "404"
        message:
          type: string
          description: Description of the error
          example: Not Found
...